Move StorageImpl to core
